Simplify update script and IGNORE handling
Dropped the notion of profiles; it doesn't really matter if some extra files
are linked on every machine. Dropped support for per-dir Makefiles, since I
wasn't using them anymore and it's cumberson. Dropped support for DOTFILES in
favor of just linking everything.

IGNORE is now a simple list of strings suitable for input to fgrep; anything
matched will not be complained about as an extra file. This makes it easier to
ignore entire directory hierarchies without having to specify each individual
file.